update tournament page url
